+++ index.js
@@ -9,14 +9,14 @@ function inject (bot) {
   bot.hawkEye.getPlayer = function (playername = null) {
     return getPlayer(bot, playername)
   }
-  bot.hawkEye.autoAttack = function (targetToAttack) {
-    return autoAttack(targetToAttack, false)
+  bot.hawkEye.autoAttack = function (targetToAttack, $weapon = 'bow') {
+    return autoAttack(targetToAttack, $weapon, false)
   }
-  bot.hawkEye.oneShot = function (targetToAttack) {
-    return autoAttack(targetToAttack, true)
+  bot.hawkEye.oneShot = function (targetToAttack, $weapon = 'bow') {
+    return autoAttack(targetToAttack, $weapon, true)
   }
-  bot.hawkEye.getMasterGrade = function (targetToAttack, speed) {
-    return getMasterGrade(bot, targetToAttack, speed)
+  bot.hawkEye.getMasterGrade = function (targetToAttack, speed, $weapon) {
+    return getMasterGrade(bot, targetToAttack, speed, $weapon)
   }
   bot.hawkEye.simplyShot = function (yaw = null, grade = null) {
     return simplyShot(bot, yaw, grade)